backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai-whisper==20231117
python-multipart==0.0.6
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1
asyncpg==0.29.0
supabase==2.1.0
PyJWT==2.8.0
langchain==0.3.26
langchain-openai==0.2.14
openai>=1.58.1

# Gmail API dependencies
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.109.0
email-validator==2.1.0

# NEW Day 1 Dependencies for Voice Pipeline + Gemma 3n
# Note: FastRTC not available, using simulation mode for now
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0
websockets>=12.0
